Update SVG generation in ClickDeduceLanguage.scala
Adjustments were made to the SVG generation code in the ClickDeduceLanguage.scala class. The modifications provide a more balanced design and size flexibility for the SVG representation of an expression tree.

Changes included a revision of the font size and calculation adjustments for the SVG size using font metrics instead of fixed pixel amounts. Logic for grouping children of the expression tree was also introduced to improve the visual layout of larger trees.